backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
mysql-connector-python
bcrypt
PyJWT
slowapi
python-dotenv
